Stop giving the kernel a framebuffer pointer so it doesnt crash on ToT magenta
We were previously passing malloc'd memory to mx_set_framebuffer, for some
reason magenta commit 797f7d43585c773326e0a52053b002f4a67fa215 made it so
this was causing our driver to fail to load. According to swetland the best
course of action is to not give the kernel an fb pointer at all.

Change-Id: Iab5da0c4289e635a6d1dbac12e389ce016ffeb24