[vkcube] Print the actual display resolution used
Before this change, vkcube prints the default resolution or the
resolution set on the command line, but not the actual resolution if
it was fetched via fpGetPhysicalDeviceSurfacePresentModesKHR().

Move the print to occur later, so that if the resolution is
successfully fetched from fpGetPhysicalDeviceSurfacePresentModesKHR(),
vkcube will print the resolution actually used.

MA-214

Change-Id: I3a86082b699d4462cd44f7adafae92faa3627086