Use GN package configurations for magma targets
This converts the magma-related targets over to using package configs
in GN instead of defining everything in JSON. This allows avoiding the
stub driver completely by defining the magma package to simply be empty
when not targeting x86-64 CPUs.

Most of the changes in build files are fixing the visibility on
intermediate GN targets so the package targets can properly express the
dependency on the files they expect to be generated, instead of relying
on luck and unrelated targets to build everything needed in the package.

Change-Id: Ie29653a2fbc80d26381af7896f4334afeef52adf